pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multilingual-sales-support-agent"
version = "0.1.0"
description = "Multilingual (EN/ES/HI) sales & support conversational AI agent with 3-agent RAG pipeline."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Aditya Srivastava" }]

dependencies = [
  # Orchestration
  "langchain==0.3.7",
  "langchain-mistralai==0.2.4",
  "langchain-community==0.3.7",
  "mistralai==1.2.5",
  # Vector store + embeddings
  "faiss-cpu==1.9.0",
  "sentence-transformers==3.3.1",
  "transformers==4.46.3",
  "torch==2.5.1",
  # API
  "fastapi==0.115.5",
  "uvicorn[standard]==0.32.1",
  "pydantic==2.10.2",
  "pydantic-settings==2.6.1",
  "python-dotenv==1.0.1",
  # Safety
  "presidio-analyzer==2.2.355",
  "presidio-anonymizer==2.2.355",
  # Language detection
  "langdetect==1.0.9",
  # Eval
  "ragas==0.2.6",
  "pandas==2.2.3",
  "scikit-learn==1.5.2",
  # Logging
  "structlog==24.4.0",
  # Retry
  "tenacity==9.0.0",
  # YAML prompts
  "pyyaml==6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-asyncio==0.24.0",
  "httpx==0.28.0",
  "ruff==0.8.1",
  "mypy==1.13.0",
  "locust==2.32.4",
]
llamaindex = ["llama-index==0.12.3", "llama-index-embeddings-huggingface==0.4.0"]
haystack   = ["haystack-ai==2.8.0"]
txtai      = ["txtai==8.1.0"]
ragatouille = ["ragatouille==0.0.9b1"]

[tool.hatch.build.targets.wheel]
packages = ["app", "ingest", "evals"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
  "integration: tests that hit live external APIs (Mistral). Gated on RUN_MISTRAL_TESTS=1.",
]
